+++ app/scripts/services/geometry/gettype.js
@@ -14,7 +14,7 @@ angular.module('accessimapEditeurDerApp')
                 switch (nodeType) {
                     case 'path':
                         var parent = feature.node().parentNode;
-                        if(d3.select(parent).classed('vector')) {
+                        if (d3.select(parent).classed('vector')) {
                             parent = parent.parentNode;
                         }
                         var parentId = parent.id;
@@ -29,7 +29,7 @@ angular.module('accessimapEditeurDerApp')
                                 return 'polygon';
                                 break;
                             default:
-                                return null;
+                                return 'polygon';
                         }
                         break;
                     case 'circle':